rename package and fix examples
